+++ Packages/Tracking/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - (Anchorable Behaviour) Code to automatically create a basic curve for attraction reach distance on new instance of the script
 - (Interaction Engine) New options to the create menu under "Interaction", can now create:
 	- Interaction Cube, 3D Button, 3D UI panel, Anchor, Anchorable Object, Attachment Hand Menu, Interaction Manager
+- (Physics Hands) Added function to check if a grasped object has been pinched
 
 ### Changed
 - (Anchorable Behaviour) Ability to change the speed at which an object is attracted to the hand